refactor(orchestrator): extract recall introspection into orchestration/recall-introspection.ts (seam 21, #1526)
Behavior-preserving move of the last-recall observability surfaces
(getLast/record/explain for intent, graph-recall, and QMD snapshots;
console faithfulness distribution; direct-answer tier annotation +
observation chain) — 616 LOC, 14 deps, one written field
(directAnswerObservationChain) via get/set accessors. parse helpers
(parseGraphRecallRankedResults, parseMemoryIntentSnapshot,
parseQmdRecallResults) exported for the coordinator.

orchestrator.ts: 7,752 -> 7,301 LOC.
